+++ simple-log/src/main/java/com/don11995/log/Extensions.kt
@@ -1,8 +1,3 @@
-/*
- * Modified by Vladyslav Lozytskyi on 05.05.18 18:46
- * Copyright (c) 2018. All rights reserved.
- */
-
 package com.don11995.log
 
 fun Class<*>.isKotlinClass(): Boolean {