+++ simple-log/src/main/java/com/don11995/log/LogLevel.kt
@@ -1,8 +1,3 @@
-/*
- * Modified by Vladyslav Lozytskyi on 04.05.18 18:06
- * Copyright (c) 2018. All rights reserved.
- */
-
 package com.don11995.log
 
 enum class LogLevel {