+++ simple-log/src/main/java/com/don11995/log/MapClassInner.kt
@@ -1,8 +1,3 @@
-/*
- * Modified by Vladyslav Lozytskyi on 04.05.18 17:39
- * Copyright (c) 2018. All rights reserved.
- */
-
 package com.don11995.log
 
 @Retention(AnnotationRetention.SOURCE)