+++ simple-log/src/main/java/com/don11995/log/MapField.kt
@@ -1,10 +1,5 @@
-/*
- * Modified by Vladyslav Lozytskyi on 05.05.18 0:30
- * Copyright (c) 2018. All rights reserved.
- */
-
 package com.don11995.log
 
 @Retention(AnnotationRetention.SOURCE)
 @Target(AnnotationTarget.FIELD)
-annotation class MapField(val value : String)
+annotation class MapField(val value: String)