+++ simple-log/src/test/java/com/don11995/log/JavaReference.java
@@ -1,8 +1,3 @@
-/*
- * Modified by Vladyslav Lozytskyi on 05.05.18 20:02
- * Copyright (c) 2018. All rights reserved.
- */
-
 package com.don11995.log;
 
 public class JavaReference {